+++ src/Dibi/Drivers/Engines/FirebirdEngine.php
@@ -7,14 +7,16 @@
 
 declare(strict_types=1);
 
-namespace Dibi\Drivers;
+namespace Dibi\Drivers\Engines;
 
+use Dibi\Drivers\Connection;
+use Dibi\Drivers\Engine;
 
 
 /**
  * The reflector for Firebird/InterBase database.
  */
-class FirebirdReflector implements Engine
+class FirebirdEngine implements Engine
 {
 	public function __construct(
 		private readonly Connection $driver,
