+++ src/Dibi/Drivers/ODBC/Connection.php
@@ -7,9 +7,10 @@
 
 declare(strict_types=1);
 
-namespace Dibi\Drivers;
+namespace Dibi\Drivers\ODBC;
 
 use Dibi;
+use Dibi\Drivers;
 use function is_resource;
 
 
@@ -24,7 +25,7 @@ use function is_resource;
  *   - resource (resource) => existing connection resource
  *   - microseconds (bool) => use microseconds in datetime format?
  */
-class OdbcDriver implements Connection
+class Connection implements Drivers\Connection
 {
 	/** @var resource */
 	private $connection;
@@ -176,9 +177,9 @@ class OdbcDriver implements Connection
 	/**
 	 * Returns the connection reflector.
 	 */
-	public function getReflector(): Engine
+	public function getReflector(): Drivers\Engine
 	{
-		return new OdbcReflector($this);
+		return new Drivers\Engines\ODBCEngine($this);
 	}
 
 
@@ -186,9 +187,9 @@ class OdbcDriver implements Connection
 	 * Result set driver factory.
 	 * @param  resource  $resource
 	 */
-	public function createResultDriver($resource): OdbcResult
+	public function createResultDriver($resource): Result
 	{
-		return new OdbcResult($resource);
+		return new Result($resource);
 	}
