+++ src/Dibi/Drivers/PgSQL/Connection.php
@@ -7,9 +7,10 @@
 
 declare(strict_types=1);
 
-namespace Dibi\Drivers;
+namespace Dibi\Drivers\PgSQL;
 
 use Dibi;
+use Dibi\Drivers;
 use Dibi\Helpers;
 use PgSql;
 use function in_array, is_array, is_resource, strlen;
@@ -27,7 +28,7 @@ use function in_array, is_array, is_resource, strlen;
  *   - resource (PgSql\Connection) => existing connection resource
  *   - connect_type (int) => see pg_connect()
  */
-class PostgreDriver implements Connection
+class Connection implements Drivers\Connection
 {
 	private PgSql\Connection $connection;
 	private ?int $affectedRows;
@@ -232,18 +233,18 @@ class PostgreDriver implements Connection
 	/**
 	 * Returns the connection reflector.
 	 */
-	public function getReflector(): Engine
+	public function getReflector(): Drivers\Engine
 	{
-		return new PostgreReflector($this);
+		return new Drivers\Engines\PostgreSQLEngine($this);
 	}
 
 
 	/**
 	 * Result set driver factory.
 	 */
-	public function createResultDriver(PgSql\Result $resource): PostgreResult
+	public function createResultDriver(PgSql\Result $resource): Result
 	{
-		return new PostgreResult($resource);
+		return new Result($resource);
 	}
