+++ src/Dibi/Drivers/SQLite3/Connection.php
@@ -7,9 +7,10 @@
 
 declare(strict_types=1);
 
-namespace Dibi\Drivers;
+namespace Dibi\Drivers\SQLite3;
 
 use Dibi;
+use Dibi\Drivers;
 use Dibi\Helpers;
 use SQLite3;
 
@@ -23,7 +24,7 @@ use SQLite3;
  *   - formatDateTime => how to format datetime in SQL (@see date)
  *   - resource (SQLite3) => existing connection resource
  */
-class SqliteDriver implements Connection
+class Connection implements Drivers\Connection
 {
 	private SQLite3 $connection;
 	private string $fmtDate;
@@ -174,18 +175,18 @@ class SqliteDriver implements Connection
 	/**
 	 * Returns the connection reflector.
 	 */
-	public function getReflector(): Engine
+	public function getReflector(): Drivers\Engine
 	{
-		return new SqliteReflector($this);
+		return new Drivers\Engines\SQLiteEngine($this);
 	}
 
 
 	/**
 	 * Result set driver factory.
 	 */
-	public function createResultDriver(\SQLite3Result $result): SqliteResult
+	public function createResultDriver(\SQLite3Result $result): Result
 	{
-		return new SqliteResult($result);
+		return new Result($result);
 	}
