+++ src/Dibi/Drivers/SQLite3/Result.php
@@ -7,31 +7,32 @@
 
 declare(strict_types=1);
 
-namespace Dibi\Drivers;
+namespace Dibi\Drivers\SQLite3;
 
 use Dibi;
+use Dibi\Drivers;
 use Dibi\Helpers;
-use PDO;
+use const SQLITE3_ASSOC, SQLITE3_BLOB, SQLITE3_FLOAT, SQLITE3_INTEGER, SQLITE3_NULL, SQLITE3_NUM, SQLITE3_TEXT;
 
 
 /**
- * The driver for PDO result set.
+ * The driver for SQLite result set.
  */
-class PdoResult implements Result
+class Result implements Drivers\Result
 {
 	public function __construct(
-		private ?\PDOStatement $resultSet,
-		private readonly string $driverName,
+		private readonly \SQLite3Result $resultSet,
 	) {
 	}
 
 
 	/**
 	 * Returns the number of rows in a result set.
+	 * @throws Dibi\NotSupportedException
 	 */
 	public function getRowCount(): int
 	{
-		return $this->resultSet->rowCount();
+		throw new Dibi\NotSupportedException('Row count is not available for unbuffered queries.');
 	}
 
 
@@ -41,12 +42,13 @@ class PdoResult implements Result
 	 */
 	public function fetch(bool $assoc): ?array
 	{
-		return Helpers::false2Null($this->resultSet->fetch($assoc ? PDO::FETCH_ASSOC : PDO::FETCH_NUM));
+		return Helpers::false2Null($this->resultSet->fetchArray($assoc ? SQLITE3_ASSOC : SQLITE3_NUM));
 	}
 
 
 	/**
 	 * Moves cursor position without fetching row.
+	 * @throws Dibi\NotSupportedException
 	 */
 	public function seek(int $row): bool
 	{
@@ -59,36 +61,24 @@ class PdoResult implements Result
 	 */
 	public function free(): void
 	{
-		$this->resultSet = null;
+		$this->resultSet->finalize();
 	}
 
 
 	/**
 	 * Returns metadata for all columns in a result set.
-	 * @throws Dibi\Exception
 	 */
 	public function getResultColumns(): array
 	{
-		$count = $this->resultSet->columnCount();
+		$count = $this->resultSet->numColumns();
 		$columns = [];
+		$types = [SQLITE3_INTEGER => 'int', SQLITE3_FLOAT => 'float', SQLITE3_TEXT => 'text', SQLITE3_BLOB => 'blob', SQLITE3_NULL => 'null'];
 		for ($i = 0; $i < $count; $i++) {
-			$row = @$this->resultSet->getColumnMeta($i); // intentionally @
-			if ($row === false) {
-				throw new Dibi\NotSupportedException('Driver does not support meta data.');
-			}
-
-			$row += [
-				'table' => null,
-				'native_type' => 'VAR_STRING',
-			];
-
 			$columns[] = [
-				'name' => $row['name'],
-				'table' => $row['table'],
-				'nativetype' => $row['native_type'],
-				'type' => $row['native_type'] === 'TIME' && $this->driverName === 'mysql' ? Dibi\Type::TimeInterval : null,
-				'fullname' => $row['table'] ? $row['table'] . '.' . $row['name'] : $row['name'],
-				'vendor' => $row,
+				'name' => $this->resultSet->columnName($i),
+				'table' => null,
+				'fullname' => $this->resultSet->columnName($i),
+				'nativetype' => $types[$this->resultSet->columnType($i)] ?? null, // buggy in PHP 7.4.4 & 7.3.16, bug 79414
 			];
 		}
 
@@ -99,7 +89,7 @@ class PdoResult implements Result
 	/**
 	 * Returns the result set resource.
 	 */
-	public function getResultResource(): ?\PDOStatement
+	public function getResultResource(): \SQLite3Result
 	{
 		return $this->resultSet;
 	}
